ua-zoom-page: Map crosshairs length nonlinearly
Because the maximum crosshair length is very long (4096 for one half, so
in total 8192 pixels), setting a relatively small crosshair is nearly
impossible with the scale, as it is extremely sensitive for small sizes.

Instead, we can map the scale value exponentially to the underlying
setting, so that at small widths, we have high precision, but at large
widths we have less precision. Using an exponential (2^x) function
"feels" good in testing.

Closes #2426